Add file via upload
A fun game I built for practice. The focus was on using functions to make my program more modular, re-writing while loops with function composition, and using more complex nested conditional statements. The interface is just the command line and at this point, the messages printed to the command line are not great, but the program functions well.